feat: add generic grid, fetch the data
